Relashionship between tables done

+++ database/migrations/2020_03_26_165226_create_users_table.php
@@ -21,7 +21,9 @@ class CreateUsersTable extends Migration
             $table->string('lastname')->length(60);
             $table->string('email')->length(100);
             $table->string('language')->length(2);
-            $table->unsignedInteger('roles_id');
+            
+            //Foreign key
+            $table->unsignedBigInteger('roles_id');
             $table->foreign('roles_id')
                 ->references('id')
                 ->on('roles')
